frontend/requirements.txt
python-fasthtml==0.10.1
httpx
uvicorn